Ported search encoding to new architecture.
git-svn-id: https://svn.apache.org/repos/asf/james/server/trunk@593739 13f79535-47bb-0310-9956-ffa450edef68

+++ imap-codec-library/src/main/java/org/apache/james/imapserver/codec/encode/ImapResponseWriter.java
@@ -50,7 +50,7 @@ public interface ImapResponseWriter {
      */
     void message( String message );
 
-    void message( int number );
+    void message( long number );
  
     /**
      * Writes a response code.
